requirements.txt
requests
msal_extensions
dnspython
